fix(docs): 用户 docs.json 索引信息覆盖问题

+++ server/config.js
@@ -62,7 +62,7 @@ const Dir = {
 
 // -- 初始化文件夹 --
 
-if (fs.existsSync(Dir.temp)) {
+if (fs.existsSync(Dir.temp) && process.env.DATA_TEST !== 'true') {
     fs.rmSync(Dir.temp, { recursive: true });
 }
